+++ lib/checkfunctions.cpp
@@ -111,7 +111,7 @@ void CheckFunctions::invalidFunctionUsage()
                 continue;
             const Token * const functionToken = tok;
             const std::vector<const Token *> arguments = getArguments(tok);
-            for (int argnr = 1; argnr <= arguments.size(); ++argnr) {
+            for (int argnr = 1; argnr <= static_cast<int>(arguments.size()); ++argnr) {
                 const Token * const argtok = arguments[argnr-1];
 
                 // check <valid>...</valid>
